change return type from int to ssize_t

+++ headers/Http.hpp
@@ -2,6 +2,7 @@
 #define HTTP_HPP
 
 #include <string>
+#include <sys/types.h>
 #include "types.hpp"
 
 // Byte stream <-> structured message, both directions; renders status codes (incl. errors).
@@ -14,7 +15,7 @@ class Http {
 		//   0: `inbuf` doesn't contain a complete request yet -- wait for more data.
 		//  -1: `inbuf` contains a malformed request -- a parse error, not a wait.
 		// Never mutates `inbuf`; the caller (Worker) owns trimming consumed bytes.
-		int          parse(const std::string &inbuf, Request &request);
+		ssize_t      parse(const std::string &inbuf, Request &request);
 		std::string  build(const Response &response);
 };
